audio: use a machine word for array sizes
we do not save anything by limiting a variable to an unsigned char,
since the compiler aligns it at machine word size anyway.  however by
using the full machine word, we save one instruction, and we remove
the useless artificial limitation to 255.

git-svn-id: https://svn.musicpd.org/mpd/trunk@7203 09075e82-0dd4-0310-85a5-a0d7c8717e4f

+++ src/audio.h
@@ -33,7 +33,7 @@ typedef struct _AudioFormat {
 	volatile mpd_sint8 bits;
 } AudioFormat;
 
-size_t audio_device_count(void);
+unsigned int audio_device_count(void);
 
 void copyAudioFormat(AudioFormat * dest, AudioFormat * src);
 
@@ -68,9 +68,9 @@ void sendMetadataToAudioDevice(MpdTag * tag);
 
 /* these functions are called in the main parent process while the child
 	process is busy playing to the audio */
-int enableAudioDevice(int fd, int device);
+int enableAudioDevice(int fd, unsigned int device);
 
-int disableAudioDevice(int fd, int device);
+int disableAudioDevice(int fd, unsigned int device);
 
 void printAudioDevices(int fd);
